chore: polish code
