+++ lampray-web/src/main/java/tech/lamprism/lampray/system/database/ssl/DatabaseSslKeyStoreArtifact.kt
@@ -16,23 +16,24 @@
 
 package tech.lamprism.lampray.system.database.ssl
 
+import java.nio.file.Path
+
 /**
- * Represents materialized SSL artifacts containing driver-specific properties
- * and resources that need cleanup.
+ * Represents a materialized SSL keystore artifact (PKCS12 or JKS).
  *
- * @param properties driver-specific SSL properties to add to the connection
- * @param resources resources to clean up when the connection is closed
+ * Contains the path to the keystore file, the password to access it,
+ * and resources for cleanup.
+ *
+ * @param path the path to the keystore file
+ * @param password the keystore password
+ * @param type the keystore type (e.g., "PKCS12", "JKS")
+ * @param resources resources to clean up the temporary keystore file
  *
  * @author RollW
  */
-data class DatabaseSslArtifacts(
-    val properties: Map<String, String> = emptyMap(),
+data class DatabaseSslKeyStoreArtifact(
+    val path: Path,
+    val password: String,
+    val type: String,
     val resources: List<AutoCloseable> = emptyList()
-) {
-    companion object {
-        /**
-         * Empty artifacts with no properties or resources.
-         */
-        val EMPTY = DatabaseSslArtifacts()
-    }
-}
+)
